lit_ecology_classifier/requirements.txt
ImageHash==4.3.1
lightning==2.2.5
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
Pillow==10.4.0
safetensors==0.4.5
scikit_learn==1.5.2
timm==0.9.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
